refactor the handler files.
